+++ src/index.ts
@@ -3,7 +3,6 @@ import { Classes as BlueprintClasses } from "@blueprintjs/core";
 import { ClassNames as IntentClassNames } from "./common/Intent";
 import * as Skeleton from "./components/Skeleton/classnames";
 import * as TypographyClassNames from "./components/Typography/classnames";
-import * as LegacyReplacements from "./legacy-replacements";
 
 const ClassNames = {
     Blueprint: BlueprintClasses,
@@ -18,4 +17,4 @@ export * from "./components";
 export * from "./extensions";
 export * from "./cmem";
 
-export { ClassNames, LegacyReplacements };
+export { ClassNames };