+++ src/legacy-replacements/index.ts
@@ -1,6 +1,0 @@
-export { ButtonReplacement as Button } from "./Button/Button";
-export { AffirmativeButtonReplacement as AffirmativeButton } from "./Button/AffirmativeButton";
-export { DismissiveButtonReplacement as DismissiveButton } from "./Button/DismissiveButton";
-export { DisruptiveButtonReplacement as DisruptiveButton } from "./Button/DisruptiveButton";
-export { RadioButtonReplacement as Radio } from "./Radio/RadioButton";
-export { TextFieldReplacement as TextField } from "./TextField/TextField";